+++ src/clj/skylobby/fx/settings.clj
@@ -21,8 +21,6 @@
 (def settings-window-width 1200)
 (def settings-window-height 1200)
 
-(def user-agent-placeholder (u/agent-string))
-
 (def default-client-id-type "random")
 (def client-id-types
   ["random"
@@ -109,6 +107,7 @@
         ready-on-unspec (fx/sub-val context :ready-on-unspec)
         ring-on-auto-unspec (fx/sub-val context :ring-on-auto-unspec)
         ring-on-spec-change (fx/sub-val context :ring-on-spec-change)
+        ring-when-game-starts (fx/sub-val context :ring-when-game-ends)
         ring-when-game-ends (fx/sub-val context :ring-when-game-ends)
         show-accolades (fx/sub-val context :show-accolades)
         show-closed-battles (fx/sub-val context :show-closed-battles)
@@ -243,6 +242,14 @@
         :selected (boolean show-closed-battles)
         :on-selected-changed {:event/type :spring-lobby/assoc
                               :key :show-closed-battles}}}
+      {:fx/type filterable-checkbox-setting
+       :title " Ring when game starts"
+       :search settings-search
+       :check-box
+       {:fx/type :check-box
+        :selected (boolean ring-when-game-starts)
+        :on-selected-changed {:event/type :spring-lobby/assoc
+                              :key :ring-when-game-starts}}}
       {:fx/type filterable-checkbox-setting
        :title " Ring when game ends"
        :search settings-search
@@ -1065,7 +1072,7 @@
           [
            {:fx/type :text-field
             :text (str user-agent-override)
-            :prompt-text (str user-agent-placeholder)
+            :prompt-text (str (u/agent-string))
             :style {:-fx-max-width 480}
             :on-text-changed {:event/type :spring-lobby/assoc
                               :key :user-agent-override}}]}