+++ src/clj/spring_lobby.clj
@@ -226,6 +226,7 @@
    :hide-passworded-battles
    :hide-spads-messages
    :hide-vote-messages
+   :hide-users-bots
    :highlight-tabs-with-new-battle-messages
    :highlight-tabs-with-new-chat-messages
    :ignore-users
@@ -259,6 +260,8 @@
    :ring-on-spec-change
    :ring-sound-file
    :ring-volume
+   :ring-when-game-starts
+   :ring-when-game-ends
    :scenarios-engine-version
    :scenarios-spring-root
    :server
@@ -394,7 +397,12 @@
 
 (def ^:dynamic *state (atom {}))
 (def ^:dynamic *ui-state
-  (atom (fx/create-context {} (cache-factory-with-threshold cache/lru-cache-factory 2048))))
+  (atom
+    (fx/create-context
+      {}
+      (cache-factory-with-threshold
+        cache/lru-cache-factory
+        1024))))
 
 (def main-stage-atom (atom nil))
 
@@ -3729,7 +3737,7 @@
 
 (defn extra-pre-game []
   (try
-    (let [{:keys [^MediaPlayer media-player music-paused]} @*state]
+    (let [{:keys [^MediaPlayer media-player music-paused ring-when-game-starts] :as state} @*state]
       (if (and media-player (not music-paused))
         (do
           (log/info "Pausing media player")
@@ -3747,7 +3755,9 @@
                   (swap! *state assoc :music-paused true))))
             (.play timeline)))
         (when (not media-player)
-          (log/info "No media player to pause"))))
+          (log/info "No media player to pause")))
+      (when ring-when-game-starts
+        (sound/play-ring state)))
     (catch Exception e
       (log/error e "Error pausing music"))))
 
@@ -3790,7 +3800,6 @@
        (fs/make-dirs spring-root)
        (catch Exception e
          (log/error e "Error creating spring root" spring-root))))
-   (alter-var-root #'skylobby.spring/extra-pre-game (constantly extra-pre-game))
    (alter-var-root #'skylobby.spring/extra-pre-game (constantly extra-pre-game))
    (alter-var-root #'skylobby.spring/extra-post-game (constantly extra-post-game))
    (alter-var-root #'skylobby.client.handler/ring-impl (constantly ring-impl))